requirements.txt
# Core
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Document Processing
python-docx>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0

# Deep Learning (for LSTM/BERT)
torch>=2.0.0
transformers>=4.30.0

# IsolationForest (from scikit-learn)
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils
numpy>=1.24.0

# Real-time Stock Data
akshare>=1.14.0

# GUI
PyQt5>=5.15.0
